Added second blog to my Appwrite series
I added a blog link in the Readme. This is a beginner-friendly blog on Medium to know more about Appwrite's functionality for cropping and optimizing images right from its API.

Links to: [Appwrite issue #1774](https://github.com/appwrite/appwrite/issues/1774)

Link to the blog: [Cropping and Optimizing images with Appwrite](https://medium.com/@anand26shweta/cropping-and-optimizing-images-with-appwrite-1e3662c37eaa)

It was very nice to get to work more with it as a beginner, I would love to know more of beginner-friendly work I can do with Appwrite as of right now. Thank you.